parsing continued

+++ android/interactable/src/main/java/com/wix/interactable/InteractableLimit.java
@@ -1,11 +1,11 @@
 package com.wix.interactable;
 
 public class InteractableLimit {
-    private double min;
-    private double max;
-    private double bounce;
+    private float min;
+    private float max;
+    private float bounce;
 
-    public InteractableLimit(double min, double max, double bounce) {
+    public InteractableLimit(float min, float max, float bounce) {
         this.min = min;
         this.max = max;
         this.bounce = bounce;
